code/Backend_Server/transfer-service/app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
